Added type cast to avoid compiler warnings about signed vs unsigned ints.

+++ src/sbml/test/TestSBMLNamespaces.cpp
@@ -235,8 +235,8 @@ START_TEST (test_SBMLNamespaces_invalid)
 {
   SBMLNamespaces *sbml = new SBMLNamespaces(3, 2);
 
-  fail_unless( sbml->getLevel() == SBML_INT_MAX );
-  fail_unless( sbml->getVersion() == SBML_INT_MAX );
+  fail_unless( sbml->getLevel() == (unsigned int) SBML_INT_MAX );
+  fail_unless( sbml->getVersion() == (unsigned int) SBML_INT_MAX );
 
   XMLNamespaces * ns = sbml->getNamespaces();
